refactor: extract createLinksFromFiles to core/utilities
